rm and fix file #72
